+++ src/main/java/com/refinedmods/refinedsites/render/release/ParsedRelease.java
@@ -2,6 +2,7 @@ package com.refinedmods.refinedsites.render.release;
 
 import com.refinedmods.refinedsites.model.release.Release;
 
+import java.util.List;
 import javax.annotation.Nullable;
 
 import lombok.Getter;
@@ -12,11 +13,12 @@ import lombok.RequiredArgsConstructor;
 public class ParsedRelease {
     private final Release release;
     private final String slug;
-    @Nullable
+    @Nullable // COMPAT
     private final String curseforgeUrl;
+    private final List<ParsedReleasePlatform> platforms;
     @Nullable
     private final String githubUrl;
-    @Nullable
+    @Nullable // COMPAT
     private final String modrinthUrl;
     @Nullable
     private final String changelogHtml;